Add media app
So far we are concentrating on music tracks.

Added Media, Tracks and IsrcNumber models. We could use another name
other than Tracks?

Not the best tests for the number generator but at this point we do not
know exactly what gets provided to us when we register as supplier, so
it will do for now.