Generate requirements.txt file from Makefile
Since the Makefile seems to be the place to do everything, add a command
to generate both the requirements.txt and requirements-dev.txt. These
can be used in place of the Pipfile. Also some external services expect
the presence of a requirements.txt and not a Pipfile.

As it stands this still means that currently the Pipfile should be where
any packages are added and then the requirements.txt is generated using
pipenv.