Add cue sheets app
This is a work in progress. The majority of functionality is there
however it is not possible to edit or remove cues from a cue sheet at
this time.

Since we are now shifting to focus more on the search functionality I
thought it best to put up what we have done so far and come back to it
later.